Improve partial replication and logging
I originally thought we might want to signal this process to request
more.  I've decided on a slow and steady approach instead.

We can see the ways we can make adjustments to the replication requests.
I don't know if requesting all logs for an author makes sense.  It may
be coming as the single-entry list.

The logging makes more sense now.  Sure to change again.