Avoid dropped packets
I wrote this trying to simplify the debug route, knowing it could drop
packets in high volume situations. Then I somehow didn't manage to fix
it until just now.

I tried this another way with `unpack_nonce_box` updating the state, but
that that just made for messier code, imo.

Not that this is unmessy, but it's fairly straight forward.