Prepare for `have` and `want` state handling
This is obviously just a step along the way, but it gets some of the
concepts in the flow.

it's also a bit less imperatively structured, which happens when I start
moving fast and breaking things

On that note, we need to get this out in more places so that we can
effecticely ask for all logs from an author.  It was specced out but
never written as part of the want message types.

So now none of the other peers understand the question yet.

In other news, I fixed the wire_buffering issue.  Essentially we stopped
generating events because they'd already filled our buffers with all of
their data.

Now we keep queueing their incoming packets until we run out.. then go
about processing them.  This area still needs work.