Use even more partial replication
Only request the new poritions of logs we have seen before. This should
include the endpoints and the bit which allow us to validate between
them... or, at least, merge them into an extant chain.

This is useful for appened, but also useful when we want to, say,
backfill after compaction.