Add some protocol documentation
If this ever becomes a thing (God forfend), I reckon it will have to be
way more specific as to "which libsodium functions do I call?" and such

As is, its a complement to the code it sits aside.

It was amazing how many times I managed to misspell receive.

Yeah, I just did it again there.