Rearrange deck chairs
Well, maybe it's not so dire.  Readers will surely note that this is not
correctly factored.  It's a step in the right direction, though.  In a
couple places we get names which make intent more clear.

On the other hand, we almost surely want an outbox queue much like our
`wire_buffer` inbox queue.  I think this simplified layout with a couple
hundred lines per file will help make reading and reasoning about
that easier.

`Protocol` is also a horrible name, but it will be changed (and
disppear) with time.