Continue refactor
Also reset the timeout for reasons which you might understand if you
were runnign a node.

This is closer to correct, but still not quite there.  I had a burst of
energy, so I got this piece done.

There's more repition in the code, by design.  I may DRY it out later,
but for now it's helping to frame my thinking.

I'm still trying not to write macros just yet.