Improve event loops
This avoids the annoying injected events (`:internal, :data`) in favor of
continuously processing the queues in essentially the order in which
they happen.

The rate limiting is not strictly necessary, but I want to be able to
have "yield" points where we can "get a message in edgewise."

This also avoids some of the calling back from `Protocol` to `Connection`
Essentially all communication in that direction is via modified state.

This, once again, includes more "DRY violations."  This continues to be
intentional.

I don't see a near-term need for parallelism.