Improve protocol definition
Using a compile time macro, we write out connection matching based on
the protocol definitions. There are other ways to accomplish this, but
this is the most extensible for our purposes.

It's not so cool to be using `next_state` when we know it should be
`keep_state` but the compiler seems to paper over this... and it
decomplicates the macro.

Extending interactions now merely means adding to the definition and
writing handlers all inside of `Protocol`.  Connection just learns about
it at compile time.