Cryout to zebra
I do not intend to keep zebra moving as quickly as quagga.  So it will
be a good regression test.  Although keeping these configs in the git
repo is a bit suboptimal.  Good enough for now.

Also includes an update to improve ~shortname matching from `Baobab`

I also wanted to see how difficult it was to set up a new instance.
Not bad.