Change timeout strategy
Instead of the annoying injected events, we just allow the timeout to
happen at regular periods. If it looks like there is still work to be
done, go round again.  Otherwise, disconnect.

This is much friendlier on human readers and actually potentially allows
for faster disconnects as the timer isn't constantly reset. It also
prevents things from piling up and not getting processed.

I believe it might also disconnect in the middle of work still if a
single operation "blocks" for a long time.  I am willing to accept this
as a sort of "the network isn't reliable" chaos monkey.