Move towards reduced replication
Start by asking for minimal items to get "up-to-date".  If we already
appear to be "up-to-date" ask for the missing pieces one-by-one.

This also eliminates (for now), special treatment of the peer.  We
eventually ask for full logs for everything.

This is a first step towards smarter replication.  It is released first
for iterative testing with live peers.

Substantive improvements to replication strategis to follow.

Some day I might even improve the undue advantages to getting a low
numbered key.