Spread log gathering work around
We've broken it up into figuring out the response set and sending each
response.  It was possible to get tied up in the combination and not be
able to respond.

I was hoping to get rid of passing the `conn_info` around, but there's
still to much state dependence.  Iterations!