Move to enacl, mostly
It turns out that there is a bug in the parameter order to the KCL auth
calls.  Not well-enough guarded to be sure.

As the only implementation (boo!) we cannot change this right away
without upgrading all of the affected bits.

I suspect I can make that happen tomorrow.  In the meantime we retain the
"broken" KCl calls for the HELLO HMACs.

All in all: buh.