Complete KCl removal (breaking change)
This fixes a bug in the `HELLO` message authentication code.
While it did not compromise security, it did do the Wrong Thing which
proper implementations would not. This was exposed by proper guard
clauses in `enacl` which prevented it from doing the Wrong Thing.

I imagine there will come a day when I add a proper guard to `KCl` to
prevent such errors, but it's not high on my list of priorities.