Change idle timeout to max connection time
Even in a situation where a peer is far behind, we want to eventually
close out any peer connection.  We can always reopen if we want to
further gossip with that peer.

This also helps with simulating bad networks even for well-connected
peers this still being obviously beta software that is OK.  Something
more robust will be desired in the future.