Add connection registry
In the long-term we want to be able to communicate with
our running processes and adjust strategies.

For now, we simply limit outbound connections to the same hort/port
pair.  There are cases where this might be an intentionally good idea.
Still, we are not going to support it right now.