Update deps
We always should have had blake2 in our own deps, but now it is missing
from the other deps

My elixir package is actually better in that it allows for the short
hashing. Which is decidely not the same as a truncated long hash.

And so it stays for our purposes.