Add replayq acceptor process
Since we can get stuff much faster than we can write, we don't
want to hold everything in memory.

replayq acts as a sort of write-ahead log.  We get the incoming data
onto the disk where we can deal with it on our own time.

In non-transient cases, this will also allow picking up where we left
off if the application closes before we finish.  Might still do
some extra transfer if we connect to another peer while things are
queued.